+++ src/mock/destination.js
@@ -1,5 +1,12 @@
 import { getRandomInteger, getRandomArrayElement } from './util.js';
-import { MIN_COUNT_DESCRIPTIONS, MAX_COUNT_DESCRIPTIONS, MIN_RANDOM_PIC, MAX_RANDOM_PIC, DESCRIPTIONS, CITIES_NAMES, MAX_PICTURES_COUNT, MIN_PICTURES_COUNT } from '../const.js';
+import { DESCRIPTIONS, CITIES_NAMES } from '../const.js';
+
+const MIN_COUNT_DESCRIPTIONS = 1;
+const MAX_COUNT_DESCRIPTIONS = 5;
+const MIN_RANDOM_PIC = 1;
+const MAX_RANDOM_PIC = 50;
+const MIN_PICTURES_COUNT = 2;
+const MAX_PICTURES_COUNT = 6;
 
 const getRandomDestinationDescription = () => {
   const count = getRandomInteger(MIN_COUNT_DESCRIPTIONS, MAX_COUNT_DESCRIPTIONS, );
