+++ src/mock/point.js
@@ -1,10 +1,15 @@
 import { getRandomInteger, getRandomArrayElement } from './util.js';
-import { POINTS_TYPES, MIN_DESTINATION_COUNT, MAX_DESTINATION_COUNT, OFFERS_BY_TYPE_COUNT, MIN_BASE_PRICE, MAX_BASE_PRICE} from '../const.js';
+import { POINTS_TYPES } from '../const.js';
 import { getRandomDates } from './dates.js';
 import { getDestination } from './destination.js';
 import { getOfferByType } from './offer.js';
 
 const MIN_ARRAY_LENGTH = 0;
+const MIN_DESTINATION_COUNT = 2;
+const MAX_DESTINATION_COUNT = 6;
+const OFFERS_BY_TYPE_COUNT = 8;
+const MIN_BASE_PRICE = 200;
+const MAX_BASE_PRICE = 2000;
 
 const offersByType = Array.from({ length: OFFERS_BY_TYPE_COUNT }, getOfferByType);
 const destinations = Array.from({ length: getRandomInteger(MIN_DESTINATION_COUNT, MAX_DESTINATION_COUNT) }, (value, index) => getDestination(index));
